+++ packages/stream_chat/test/src/mocks.dart
@@ -1,7 +1,6 @@
 import 'package:dio/dio.dart';
 import 'package:logging/logging.dart';
 import 'package:mocktail/mocktail.dart';
-import 'package:rxdart/rxdart.dart';
 import 'package:stream_chat/src/client/channel.dart';
 import 'package:stream_chat/src/client/client.dart';
 import 'package:stream_chat/src/core/api/attachment_file_uploader.dart';
@@ -18,6 +17,7 @@ import 'package:stream_chat/src/core/http/stream_http_client.dart';
 import 'package:stream_chat/src/core/http/token_manager.dart';
 import 'package:stream_chat/src/core/models/channel_config.dart';
 import 'package:stream_chat/src/core/models/event.dart';
+import 'package:stream_chat/src/core/util/event_controller.dart';
 import 'package:stream_chat/src/db/chat_persistence_client.dart';
 import 'package:stream_chat/src/event_type.dart';
 import 'package:stream_chat/src/ws/websocket.dart';
@@ -98,7 +98,7 @@ class MockStreamChatClient extends Mock implements StreamChatClient {
 
   @override
   Stream<Event> get eventStream => _eventController.stream;
-  final _eventController = PublishSubject<Event>();
+  final _eventController = EventController<Event>();
   void addEvent(Event event) => _eventController.add(event);
 
   @override