+++ packages/stream_chat_persistence/lib/src/dao/pinned_message_dao.dart
@@ -38,6 +38,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
   Future<Message> _messageFromJoinRow(
     TypedResult rows, {
     bool fetchDraft = false,
+    bool fetchSharedLocation = false,
   }) async {
     final userEntity = rows.readTableOrNull(_users);
     final pinnedByEntity = rows.readTableOrNull(_pinnedByUsers);
@@ -68,6 +69,11 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
       _ => null,
     };
 
+    final sharedLocation = await switch (fetchSharedLocation) {
+      true => _db.locationDao.getLocationByMessageId(msgEntity.id),
+      _ => null,
+    };
+
     return msgEntity.toMessage(
       user: userEntity?.toUser(),
       pinnedBy: pinnedByEntity?.toUser(),
@@ -76,6 +82,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
       quotedMessage: quotedMessage,
       poll: poll,
       draft: draft,
+      sharedLocation: sharedLocation,
     );
   }
 
@@ -83,6 +90,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
   Future<Message?> getMessageById(
     String id, {
     bool fetchDraft = true,
+    bool fetchSharedLocation = true,
   }) async {
     final query = select(pinnedMessages).join([
       leftOuterJoin(_users, pinnedMessages.userId.equalsExp(_users.id)),
@@ -99,6 +107,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
     return _messageFromJoinRow(
       result,
       fetchDraft: fetchDraft,
+      fetchSharedLocation: fetchSharedLocation,
     );
   }
 
@@ -166,6 +175,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
   Future<List<Message>> getMessagesByCid(
     String cid, {
     bool fetchDraft = true,
+    bool fetchSharedLocation = true,
     PaginationParams? messagePagination,
   }) async {
     final query = select(pinnedMessages).join([
@@ -188,6 +198,7 @@ class PinnedMessageDao extends DatabaseAccessor<DriftChatDatabase>
         (row) => _messageFromJoinRow(
           row,
           fetchDraft: fetchDraft,
+          fetchSharedLocation: fetchSharedLocation,
         ),
       ),
     );